[refactor] ProfileCallback과 OAuthLoginCallback 객체 병합
1. 서버와 통신할 때에는 OAuth Access Token만이 필요하기 때문에 두 클래스를 병합하였다.

Close: #41
